+++ tkcalendar/dateentry.py
@@ -338,6 +338,8 @@ class DateEntry(ttk.Entry):
             entry_kw['font'] = font
         ttk.Entry.configure(self, **entry_kw)
         self._calendar.configure(**kw)
+        if 'date_pattern' in kw or 'locale' in kw:
+            self._set_text(self.format_date(self._date))
 
     config = configure