+++ docs/changelog.rst
@@ -9,6 +9,7 @@ tkcalendar 1.5.0
 .. rubric:: New features
 
 - :meth:`Calendar.see` method: make sure given date is visible
+- Make *locale* option editable after the creation of the :class:`Calendar`
 
 .. rubric:: New options
 
@@ -16,6 +17,7 @@ tkcalendar 1.5.0
   month/year name in disabled state
 - *maxdate* and *mindate*: set an allowed date range for date selection
 - *weekenddays*: choose the days colored as week-end days (`#37 <https://github.com/j4321/tkcalendar/issues/37>`_)
+- *date_pattern*: customize the date format
 
 .. rubric:: Bug fixes